requirements.txt
Django==6.0.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0
python-decouple==3.8
mysqlclient==2.2.4
google-generativeai==0.8.6
firebase_admin==7.3.0
requests==2.32.5
PyPDF2>=3.0.1
pillow==12.1.1
